// src/app/interfaces/user-model.ts
export interface UserModel {
    uID?: string;
    uid?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    phone?: string;
    address?: string;
    displayName?: string;
    birthDate?: string;
    bloodGroup?: string;
    studyInfo?: string;
    photoURL?: string;
    coverImageHD?: string;
    imageThumbnail?: string;
    registrationType?: string;
    role: string;
    registeredTime: any;
}